chore(tests): move unit tests to a subfolder
