+++ tests/Unit/Document/CertificateTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Unikorp\KongAdminApi\Tests\Document;
+namespace Unikorp\KongAdminApi\Tests\Unit\Document;
 
 use Unikorp\KongAdminApi\Document\Certificate as Document;
 use PHPUnit\Framework\TestCase;
