+++ tests/Unit/Document/TargetTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Unikorp\KongAdminApi\Tests\Document;
+namespace Unikorp\KongAdminApi\Tests\Unit\Document;
 
 use Unikorp\KongAdminApi\Document\Target as Document;
 use PHPUnit\Framework\TestCase;
